add player_CCSystemEvent for fireball/issues/4237
