+++ jsb/jsb-predefine.js
@@ -54,6 +54,7 @@ require('../cocos2d/core/value-types');
 require('../cocos2d/core/utils/find');
 require('../cocos2d/core/utils/mutable-forward-iterator');
 require('../cocos2d/core/event');
+require('../cocos2d/core/event-manager/CCSystemEvent');
 require('../CCDebugger');
 
 cc._initDebugSetting(cc.game.DEBUG_MODE_INFO);
